Move histogram and cumulative_histogram from contrast module to stats module
